Refactor radio button row component
