+++ src/components/SharedComponents/index.js
@@ -32,6 +32,7 @@ export { default as OverviewCounts } from "./OverviewCounts";
 export { default as PermissionGate } from "./PermissionGate";
 export { default as PhotoCount } from "./PhotoCount";
 export { default as QualityGradeStatus } from "./QualityGradeStatus/QualityGradeStatus";
+export { default as RadioButtonRow } from "./RadioButtonRow";
 export { default as ScrollViewWrapper } from "./ScrollViewWrapper";
 export { default as SearchBar } from "./SearchBar";
 export { default as BottomSheet } from "./Sheets/BottomSheet";
